+++ src/java/iced-x86/src/main/java/com/github/icedland/iced/x86/fmt/masm/MasmFormatter.java
@@ -635,7 +635,7 @@ public final class MasmFormatter extends Formatter {
 			}
 			else {
 				if (numberOptions.signedNumber) {
-					imm64 = (byte)imm8;
+					imm64 = imm8;
 					numberKind = NumberKind.INT8;
 					if (imm8 < 0) {
 						output.write("-", FormatterTextKind.OPERATOR);
@@ -674,7 +674,7 @@ public final class MasmFormatter extends Formatter {
 			}
 			else {
 				if (numberOptions.signedNumber) {
-					imm64 = (short)imm16;
+					imm64 = imm16;
 					numberKind = NumberKind.INT16;
 					if (imm16 < 0) {
 						output.write("-", FormatterTextKind.OPERATOR);
